+++ hutool-setting/src/main/java/cn/hutool/setting/dialect/Props.java
@@ -21,6 +21,7 @@ import cn.hutool.core.util.CharsetUtil;
 import cn.hutool.core.util.ReflectUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.log.StaticLog;
+import cn.hutool.setting.Setting;
 import cn.hutool.setting.SettingRuntimeException;
 
 import java.io.BufferedReader;
@@ -49,6 +50,16 @@ public final class Props extends Properties implements BasicTypeGetter<String>, 
 	 * 默认配置文件扩展名
 	 */
 	public final static String EXT_NAME = "properties";
+
+	/**
+	 * 构建一个空的Props，用于手动加入参数
+	 *
+	 * @return Setting
+	 * @since 5.4.3
+	 */
+	public static Props create() {
+		return new Props();
+	}
 
 	// ----------------------------------------------------------------------- 私有属性 start
 	/** 属性文件的URL */